+++ src/main/java/classfit/example/classfit/auth/security/custom/CustomUserDetailService.java
@@ -1,4 +1,4 @@
-package classfit.example.classfit.auth.security.provider;
+package classfit.example.classfit.auth.security.custom;
 
 import classfit.example.classfit.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
